Separated Card components into multiple components
separated inner cop card components into its own file js and scss files.
