+++ app/frontend/src/components/Cards/_Cards.scss
@@ -75,38 +75,3 @@
     left: 6.4vw;
   }
 }
-
-/**********************************
-*** Classes for inner COP cards ***
-**********************************/
-.inner-cop-card {
-  background-color: $color-white;
-  box-shadow: 0 8px 8px rgb(0 0 0 / 25%);
-  max-width: 625px;
-  min-height: 448px;
-}
-
-.inner-cop-card-content {
-  padding: {
-    top: 48px;
-    right: 3.3vw;
-    left: 3.3vw;
-  }
-}
-
-/**************************************
-*** Classes for inner COP nav cards ***
-**************************************/
-.inner-cop-nav-card {
-  background-color: $color-white;
-  border: none;
-  box-sizing: border-box;
-  display: inline;
-  max-width: 188px;
-  min-height: 64px;
-
-  &.active {
-    background-color: $color-blue-dark;
-    color: $color-white;
-  }
-}
